rename the KeysightE3631A.py to keysightE3631A.py
