+++ pymeasure/instruments/keysight/__init__.py
@@ -26,4 +26,4 @@ from .keysightDSOX1102G import KeysightDSOX1102G
 from .keysightN5767A import KeysightN5767A
 from .keysightN7776C import KeysightN7776C
 from .keysightE36312A import KeysightE36312A
-from .KeysightE3631A import KeysightE3631A
+from .keysightE3631A import KeysightE3631A
